refactor: extract call tx flow
